2 buttons work and RunAfter
The start button works and brings you to level 1, The Level selection button works but the page is empty at this point.

RunAfter is used to run certain methodes a bit later. Like enabling the Buttons in the UI a bit later so it looks better with the Page animation